# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/accounts": {
      "get": {
        "operationId": "AccountsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all accounts",
        "tags": [
          "accounts"
        ]
      },
      "post": {
        "operationId": "AccountsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new account",
        "tags": [
          "accounts"
        ]
      }
    },
...
    "/accounts/{id}": {
      "get": {
        "operationId": "AccountsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get account by it's id",
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "operationId": "AccountsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update account",
        "tags": [
          "accounts"
        ]
      },
      "delete": {
        "operationId": "AccountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Account removed"
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove account",
        "tags": [
          "accounts"
        ]
      }
    },
...
    "/solana-dca/accounts": {
      "get": {
        "description": "This method returns all solana DCA accounts than created and assigned to the accounts. You can explitly edit some of them",
        "operationId": "SolanaDcaController_getAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of solana DCA accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DcaAccountEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all solana DCA accounts",
        "tags": [
          "solana-dca"
        ]
      }
    },
    "/solana-dca/accounts/all": {
      "patch": {
        "description": "Updates all existing solana DCA accounts with provided data",
        "operationId": "SolanaDcaController_updateAllAccounts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDcaAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated solana DCA accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DcaAccountEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update all solana DCA accounts",
        "tags": [
          "solana-dca"
        ]
      }
    },
    "/solana-dca/accounts/{id}": {
      "patch": {
        "description": "This method updates solana DCA account",
        "operationId": "SolanaDcaController_updateAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDcaAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated solana DCA account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DcaAccountEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update solana DCA account",
        "tags": [
          "solana-dca"
        ]
      }
    },
    "/solana-dca/on": {
      "post": {
        "description": "This method activates all solana DCA accounts",
        "operationId": "SolanaDcaController_turnOnAllAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Activated solana DCA accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DcaAccountEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate all solana DCA accounts",
        "tags": [
          "solana-dca"
        ]
      }
    },
    "/solana-dca/off": {
      "delete": {
        "description": "This method deactivates all solana DCA accounts",
        "operationId": "SolanaDcaController_turnOffAllAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Deactivated solana DCA accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DcaAccountEntity"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate all solana DCA accounts",
        "tags": [
          "solana-dca"
        ]
      }
    }
  },
  "info": {
    "title": "Bundler API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AccountEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the account",
            "example": "c24ef4b2-bd4a-4d5a-ae4c-d1b3ee1e4f4e"
          },
          "type": {
            "type": "string",
            "description": "The type of the account",
            "example": "SOLANA"
          },
          "userId": {
            "type": "string",
            "description": "The unique identifier of the user that created the account",
            "example": "c24ef4b2-bd4a-4d5a-ae4c-d1b3ee1e4f4e"
          },
          "color": {
            "type": "string",
            "description": "The color of the account",
            "example": "#ff0000"
          },
          "name": {
            "type": "string",
            "description": "The name of the account",
            "example": "JohnDoe"
          },
          "publicKey": {
            "type": "string",
            "description": "The public key of the solana account",
            "example": "5HpHagT65TZzG1PH3CSu63k8DbpvD8s5ip4nEB3kEsreAnchuDf"
          },
          "tokenAccountId": {
            "type": "object",
            "description": "The public key of the solana token account",
            "example": "5HpHagT65TZzG1PH3CSu63k8DbpvD8s5ip4nEB3kEsreAnchuDf"
          },
          "balance": {
            "type": "number",
            "description": "The balance of the account",
            "example": 5.123456789
          },
          "tokenBalance": {
            "type": "number",
            "description": "The token balance of the account",
            "example": 124532
          },
          "isBalanceSynced": {
            "type": "boolean",
            "description": "Whether the account balance is synced or not"
          },
          "isTokenBalanceSynced": {
            "type": "boolean",
            "description": "Whether the account token balance is synced or not"
          },
          "isTokenAccountInitialized": {
            "type": "boolean",
            "description": "Whether the account token account is initialized or not"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the account is active or not"
          },
          "isImported": {
            "type": "boolean",
            "description": "Whether the account is imported or not"
          },
          "isRemoved": {
            "type": "boolean",
            "description": "Whether the account is removed or not"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date when the account was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date when the account was last updated"
          },
          "syncedAt": {
            "type": "object",
            "description": "The date when the account was synced"
          },
          "syncProblemInspectedAt": {
            "type": "object",
            "description": "System timestamp that indicates is there any problem with the account and syncing"
          },
          "lastBuyAt": {
            "type": "object",
            "description": "The date when the account was last bought"
          },
          "lastSellAt": {
            "type": "object",
            "description": "The date when the account was last sold"
          },
          "removedAt": {
            "type": "object",
            "description": "The date when the account was removed"
          },
          "exportedAt": {
            "type": "object",
            "description": "The date when the account was exported"
          }
        },
        "required": [
          "id",
          "type",
          "publicKey",
          "tokenAccountId",
          "balance",
          "tokenBalance",
          "isBalanceSynced",
          "isTokenBalanceSynced",
          "isTokenAccountInitialized",
          "isActive",
          "isImported",
          "isRemoved",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string",
            "description": "The color of the account",
            "example": "#ff0000"
          },
          "name": {
            "type": "string",
            "description": "The name of the account",
            "example": "JohnDoe"
          }
        }
      },
...
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "string",
            "description": "The color of the account",
            "example": "#ff0000"
          },
          "name": {
            "type": "string",
            "description": "The name of the account",
            "example": "JohnDoe"
          }
        }
      },
...
      "DcaAccountEntity": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "The id of the solana DCA account (also public key)"
          },
          "bumpOperateSolAmount": {
            "type": "number",
            "description": "The amount of SOL that will be used for each transaction",
            "example": 0.01
          },
          "balanceUsagePercent": {
            "type": "number",
            "description": "How much tokens we will sell every time",
            "example": 50
          },
          "minTokenPrice": {
            "type": "number",
            "description": "The minimum price of the token to sell",
            "example": 2.78e-8
          },
          "maxTokenPrice": {
            "type": "number",
            "description": "The maximum price of the token to sell",
            "example": 0.000008
          },
          "minDelayBetweenTxsInSeconds": {
            "type": "number",
            "description": "The minimum delay between transactions for account in seconds",
            "example": 15
          },
          "maxDelayBetweenTxsInSeconds": {
            "type": "number",
            "description": "The maximum delay between transactions for account in seconds",
            "example": 60
          },
          "slippagePercent": {
            "type": "number",
            "description": "How much we ready to lost",
            "example": 5
          },
          "reserveSolAmount": {
            "type": "number",
            "description": "How much solana we will store in any circumstance",
            "example": 0.1
          },
          "minTokenAmountPerSale": {
            "type": "number",
            "description": "When we will sell the token, we need to have at least this amount",
            "example": 1000
          },
          "maxTokenAmount": {
            "type": "number",
            "description": "How much tokens we can buy",
            "example": 20000000
          },
          "reserveTokenAmount": {
            "type": "number",
            "description": "How much tokens we will store in any circumstance",
            "example": 2000000
          },
          "canBuy": {
            "type": "boolean",
            "description": "Is buying allowed for auto orders?",
            "example": true
          },
          "canSell": {
            "type": "boolean",
            "description": "Is selling allowed for auto orders?",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "description": "When active transactions will be working",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date when the account was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date when the account was last updated"
          },
          "lastTxType": {
            "type": "object",
            "description": "Last tx type"
          },
          "allowNextAt": {
            "format": "date-time",
            "type": "string",
            "description": "Allow next at date"
          }
        },
        "required": [
          "accountId",
          "bumpOperateSolAmount",
          "balanceUsagePercent",
          "minTokenPrice",
          "maxTokenPrice",
          "minDelayBetweenTxsInSeconds",
          "maxDelayBetweenTxsInSeconds",
          "slippagePercent",
          "reserveSolAmount",
          "minTokenAmountPerSale",
          "maxTokenAmount",
          "reserveTokenAmount",
          "canBuy",
          "canSell",
          "isActive",
          "createdAt",
          "updatedAt",
          "allowNextAt"
        ]
      },
      "UpdateDcaAccountDto": {
        "type": "object",
        "properties": {
          "bumpOperateSolAmount": {
            "type": "number",
            "description": "The amount of SOL that will be used for each transaction",
            "example": 0.01
          },
          "balanceUsagePercent": {
            "type": "number",
            "description": "How much tokens we will sell every time",
            "example": 50
          },
          "minTokenPrice": {
            "type": "number",
            "description": "The minimum price of the token to sell",
            "example": 2.78e-8
          },
          "maxTokenPrice": {
            "type": "number",
            "description": "The maximum price of the token to sell",
            "example": 0.000008
          },
          "minDelayBetweenTxsInSeconds": {
            "type": "number",
            "description": "The minimum delay between transactions for account in seconds",
            "example": 15
          },
          "maxDelayBetweenTxsInSeconds": {
            "type": "number",
            "description": "The maximum delay between transactions for account in seconds",
            "example": 60
          },
          "slippagePercent": {
            "type": "number",
            "description": "How much we ready to lost",
            "example": 5
          },
          "reserveSolAmount": {
            "type": "number",
            "description": "How much solana we will store in any circumstance",
            "example": 0.1
          },
          "minTokenAmountPerSale": {
            "type": "number",
            "description": "When we will sell the token, we need to have at least this amount",
            "example": 1000
          },
          "maxTokenAmount": {
            "type": "number",
            "description": "How much tokens we can buy",
            "example": 20000000
          },
          "reserveTokenAmount": {
            "type": "number",
            "description": "How much tokens we will store in any circumstance",
            "example": 2000000
          },
          "canBuy": {
            "type": "boolean",
            "description": "Is buying allowed for auto orders?",
            "example": true
          },
          "canSell": {
            "type": "boolean",
            "description": "Is selling allowed for auto orders?",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "description": "When active transactions will be working",
            "example": true
          },
          "lastTxType": {
            "type": "object",
            "description": "Last tx type"
          },
          "allowNextAt": {
            "format": "date-time",
            "type": "string",
            "description": "Allow next at date"
          }
        }
      }
    }
  }
}